[UI] activty_set_profile FramLayout 으로변경 , fragment_set_profile 추가 및 다음 화면으로 넘어가는 로직 추가
